fix: remove the highlighted nodes #374
